add internal state and stop
