+++ adapters/memrecordstore/memrecordstore.go
