+++ adapters/sqlstore/util.go
