+++ adapters/sqlstore/util_test.go
