+++ adapters/testing/recordstore.go
