+++ await.go
@@ -1,0 +1,33 @@
+package workflow
+
+import (
+	"context"
+	"time"
+)
+
+func (w *Workflow[Type, Status]) Await(ctx context.Context, foreignID, runID string, status Status, opts ...AwaitOption) (*Record[Type, Status], error) {
+	var opt awaitOpts
+	for _, option := range opts {
+		option(&opt)
+	}
+
+	pollFrequency := w.defaultPollingFrequency
+	if opt.pollFrequency.Nanoseconds() != 0 {
+		pollFrequency = opt.pollFrequency
+	}
+
+	role := makeRole("await", w.Name, string(status), foreignID)
+	return awaitWorkflowStatusByForeignID[Type, Status](ctx, w, status, foreignID, runID, role, pollFrequency)
+}
+
+type awaitOpts struct {
+	pollFrequency time.Duration
+}
+
+type AwaitOption func(o *awaitOpts)
+
+func WithPollingFrequency(d time.Duration) AwaitOption {
+	return func(o *awaitOpts) {
+		o.pollFrequency = d
+	}
+}
