+++ callback.go
@@ -9,6 +9,24 @@ import (
 	"github.com/luno/jettison/j"
 )
 
+type callback[Type any, Status ~string] struct {
+	DestinationStatus Status
+	CallbackFunc      CallbackFunc[Type, Status]
+}
+
+type CallbackFunc[Type any, Status ~string] func(ctx context.Context, r *Record[Type, Status], reader io.Reader) (bool, error)
+
+func (w *Workflow[Type, Status]) Callback(ctx context.Context, foreignID string, status Status, payload io.Reader) error {
+	for _, s := range w.callback[status] {
+		err := processCallback(ctx, w, status, s.DestinationStatus, s.CallbackFunc, foreignID, payload)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
 func processCallback[Type any, Status ~string](ctx context.Context, w *Workflow[Type, Status], currentStatus, destinationStatus Status, fn CallbackFunc[Type, Status], foreignID string, payload io.Reader) error {
 	latest, err := w.recordStore.Latest(ctx, w.Name, foreignID)
 	if err != nil {
@@ -18,7 +36,7 @@ func processCallback[Type any, Status ~string](ctx context.Context, w *Workflow[
 	}
 
 	if latest.Status != string(currentStatus) {
-		// Latest record shows that the current status is in a different state than expected so skip.
+		// Latest record shows that the current status is in a different State than expected so skip.
 		return nil
 	}
