--- connectors/ui/public/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Workflow</title>
-</head>
-<body>
-    <div id="root"></div>
-    <script src="index.js"></script>
-</body>
-</html>
