--- connectors/ui/src/App.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import BasicTable from "./Runs";
-
-function App() {
-    return (
-        <div>
-            <h1>Workflow</h1>
-            <div className="table-wrapper">
-                <BasicTable/>
-            </div>
-        </div>
-    );
-}
-
-
-
-export default App;
