--- connectors/ui/src/styles.css
@@ -1,13 +1,0 @@
-.table-wrapper {
-    padding: 15px;
-    display: flex;
-    justify-content: center;
-}
-
-.is-complete {
-    background: forestgreen;
-}
-
-.is-pending {
-    background: orange;
-}
