+++ producer.go
@@ -1,8 +1,2 @@
 package workflow
 
-import "context"
-
-type Producer interface {
-	Send(ctx context.Context, e *Event) error
-	Close() error
-}
