+++ workflow.go
@@ -2,22 +2,45 @@ package workflow
 
 import (
 	"context"
-	"fmt"
-	"github.com/robfig/cron/v3"
 	"io"
-	"strings"
 	"sync"
 	"time"
 
-	"github.com/google/uuid"
-	"github.com/luno/jettison/errors"
-	"github.com/luno/jettison/j"
-	"github.com/luno/jettison/log"
 	"k8s.io/utils/clock"
 )
 
+type Protocol[Type any, Status ~string] interface {
+	// Trigger will kickstart a workflow for the provided foreignID starting from the provided starting status. There
+	// is no limitation as to where you start the workflow from. For workflows that have data preceding the initial
+	// trigger that needs to be used in the workflow, using WithInitialValue will allow you to provide pre-populated
+	// fields of Type that can be accessed by the consumers.
+	Trigger(ctx context.Context, foreignID string, startingStatus Status, opts ...TriggerOption[Type, Status]) (runID string, err error)
+
+	// ScheduleTrigger takes a cron spec and will call Trigger at the specified intervals. The same options are
+	// available for ScheduleTrigger than they are for Trigger.
+	ScheduleTrigger(ctx context.Context, foreignID string, startingStatus Status, spec string, opts ...TriggerOption[Type, Status]) error
+
+	// Await is a blocking call that returns the typed Record when the workflow of the specified run ID reaches the
+	// specified status.
+	Await(ctx context.Context, foreignID, runID string, status Status, opts ...AwaitOption) (*Record[Type, Status], error)
+
+	// Callback can be used if Builder.AddCallback has been defined for the provided status. The data in the reader
+	// will be passed to the CallbackFunc that you specify and so the serialisation and deserialisation is in the
+	// hands of the user.
+	Callback(ctx context.Context, foreignID string, status Status, payload io.Reader) error
+
+	// Run must be called in order to start up all the background consumers / consumers required to run the workflow. Run
+	// only needs to be called once. Any subsequent calls to run are safe and are noop.
+	Run(ctx context.Context)
+
+	// Stop tells the workflow to shutdown gracefully.
+	Stop()
+}
+
 type Workflow[Type any, Status ~string] struct {
 	Name string
+
+	cancel context.CancelFunc
 
 	clock                   clock.Clock
 	defaultPollingFrequency time.Duration
@@ -26,14 +49,19 @@ type Workflow[Type any, Status ~string] struct {
 	calledRun bool
 	once      sync.Once
 
-	eventStreamerFn EventStreamerConstructor
+	eventStreamerFn EventStreamer
 	recordStore     RecordStore
 	timeoutStore    TimeoutStore
 	scheduler       RoleScheduler
 
-	processes map[Status][]process[Type, Status]
+	consumers map[Status][]consumerConfig[Type, Status]
 	callback  map[Status][]callback[Type, Status]
 	timeouts  map[Status]timeouts[Type, Status]
+
+	internalStateMu sync.Mutex
+	// internalState holds the State of all expected consumers and timeout  go routines using their role names
+	// as the key.
+	internalState map[string]State
 
 	graph         map[Status][]Status
 	endPoints     map[Status]bool
@@ -42,224 +70,17 @@ type Workflow[Type any, Status ~string] struct {
 	debugMode bool
 }
 
-func (w *Workflow[Type, Status]) Trigger(ctx context.Context, foreignID string, startingStatus Status, opts ...TriggerOption[Type, Status]) (runID string, err error) {
-	if !w.calledRun {
-		return "", errors.Wrap(ErrWorkflowNotRunning, "ensure Run() is called before attempting to trigger the workflow")
-	}
-
-	_, ok := w.validStatuses[startingStatus]
-	if !ok {
-		return "", errors.Wrap(ErrStatusProvidedNotConfigured, fmt.Sprintf("ensure %v is configured for workflow: %v", startingStatus, w.Name))
-	}
-
-	var o triggerOpts[Type, Status]
-	for _, fn := range opts {
-		fn(&o)
-	}
-
-	var t Type
-	if o.initialValue != nil {
-		t = *o.initialValue
-	}
-
-	object, err := Marshal(&t)
-	if err != nil {
-		return "", err
-	}
-
-	lastRecord, err := w.recordStore.Latest(ctx, w.Name, foreignID)
-	if errors.Is(err, ErrRecordNotFound) {
-		lastRecord = &WireRecord{}
-	} else if err != nil {
-		return "", err
-	}
-
-	// Check that the last entry for that workflow was a terminal step when entered.
-	if lastRecord.RunID != "" && !lastRecord.IsEnd {
-		// Cannot trigger a new workflow for this foreignID if there is a workflow in progress
-		return "", errors.Wrap(ErrWorkflowInProgress, "")
-	}
-
-	uid, err := uuid.NewUUID()
-	if err != nil {
-		return "", err
-	}
-
-	runID = uid.String()
-	wr := &WireRecord{
-		RunID:        runID,
-		WorkflowName: w.Name,
-		ForeignID:    foreignID,
-		Status:       string(startingStatus),
-		// isStart is always true when being stored as the trigger as it is the beginning of the workflow
-		IsStart: true,
-		// isEnd is always false as there should always be more than one node in the graph so that there can be a
-		// transition between statuses / states.
-		IsEnd:     false,
-		Object:    object,
-		CreatedAt: w.clock.Now(),
-	}
-
-	err = update(ctx, w.eventStreamerFn, w.recordStore, wr)
-	if err != nil {
-		return "", err
-	}
-
-	return runID, nil
-}
-
-func (w *Workflow[Type, Status]) ScheduleTrigger(ctx context.Context, foreignID string, startingStatus Status, spec string, opts ...TriggerOption[Type, Status]) error {
-	if !w.calledRun {
-		return errors.Wrap(ErrWorkflowNotRunning, "ensure Run() is called before attempting to trigger the workflow")
-	}
-
-	_, ok := w.validStatuses[startingStatus]
-	if !ok {
-		return errors.Wrap(ErrStatusProvidedNotConfigured, fmt.Sprintf("ensure %v is configured for workflow: %v", startingStatus, w.Name))
-	}
-
-	schedule, err := cron.ParseStandard(spec)
-	if err != nil {
-		return err
-	}
-
-	role := strings.Join([]string{w.Name, string(startingStatus), foreignID, "scheduler", spec}, "-")
-
-	for {
-		ctx, cancel, err := w.scheduler.Await(ctx, role)
-		if err != nil {
-			log.Error(ctx, errors.Wrap(err, "timeout auto inserter consumer error"))
-		}
-
-		latestEntry, err := w.recordStore.Latest(ctx, w.Name, foreignID)
-		if errors.Is(err, ErrRecordNotFound) {
-			// NoReturnErr: Rather use zero value for lastRunID and use current clock for first run.
-			latestEntry = &WireRecord{}
-		} else if err != nil {
-			log.Error(ctx, errors.Wrap(err, "schedule trigger error - lookup latest record", j.MKV{
-				"workflow_name":   w.Name,
-				"foreignID":       foreignID,
-				"starting_status": startingStatus,
-			}))
-			cancel()
-			continue
-		}
-
-		lastRun := latestEntry.CreatedAt
-
-		// If there is no previous executions of this workflow then schedule the very next from now.
-		if lastRun.IsZero() {
-			lastRun = w.clock.Now()
-		}
-
-		nextRun := schedule.Next(lastRun)
-		err = waitUntil(ctx, w.clock, nextRun)
-		if errors.Is(err, context.Canceled) {
-			cancel()
-			continue
-		} else if err != nil {
-			log.Error(ctx, errors.Wrap(err, "schedule trigger error - wait until", j.MKV{
-				"workflow_name": w.Name,
-				"now":           w.clock.Now(),
-				"spec":          spec,
-				"last_run":      lastRun,
-				"next_run":      nextRun,
-			}))
-			cancel()
-			continue
-		}
-
-		_, err = w.Trigger(ctx, foreignID, startingStatus, opts...)
-		if errors.Is(err, ErrWorkflowInProgress) {
-			// NoReturnErr: Fallthrough to schedule next workflow as there is already one in progress. If this
-			// happens it is likely that we scheduled a workflow and were unable to schedule the next.
-		} else if err != nil {
-			log.Error(ctx, errors.Wrap(err, "schedule trigger error - triggering workflow", j.MKV{
-				"workflow_name":   w.Name,
-				"foreignID":       foreignID,
-				"starting_status": startingStatus,
-			}))
-			cancel()
-			continue
-		}
-
-		// Always cancel the context to release the role
-		cancel()
-	}
-}
-
-func waitUntil(ctx context.Context, clock clock.Clock, until time.Time) error {
-	timeDiffAsDuration := until.Sub(clock.Now())
-
-	t := clock.NewTimer(timeDiffAsDuration)
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	case <-t.C():
-		return nil
-	}
-}
-
-type triggerOpts[Type any, Status ~string] struct {
-	initialValue   *Type
-	startingStatus int
-}
-
-type TriggerOption[Type any, Status ~string] func(o *triggerOpts[Type, Status])
-
-func WithInitialValue[Type any, Status ~string](t *Type) TriggerOption[Type, Status] {
-	return func(o *triggerOpts[Type, Status]) {
-		o.initialValue = t
-	}
-}
-
-func (w *Workflow[Type, Status]) Await(ctx context.Context, foreignID, runID string, status Status, opts ...AwaitOption) (*Record[Type, Status], error) {
-	var opt awaitOpts
-	for _, option := range opts {
-		option(&opt)
-	}
-
-	pollFrequency := w.defaultPollingFrequency
-	if opt.pollFrequency.Nanoseconds() != 0 {
-		pollFrequency = opt.pollFrequency
-	}
-
-	role := makeRole("await", w.Name, string(status), foreignID)
-	return awaitWorkflowStatusByForeignID[Type, Status](ctx, w, status, foreignID, runID, role, pollFrequency)
-}
-
-type awaitOpts struct {
-	pollFrequency time.Duration
-}
-
-type AwaitOption func(o *awaitOpts)
-
-func WithPollingFrequency(d time.Duration) AwaitOption {
-	return func(o *awaitOpts) {
-		o.pollFrequency = d
-	}
-}
-
-func (w *Workflow[Type, Status]) Callback(ctx context.Context, foreignID string, status Status, payload io.Reader) error {
-	for _, s := range w.callback[status] {
-		err := processCallback(ctx, w, status, s.DestinationStatus, s.CallbackFunc, foreignID, payload)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
 func (w *Workflow[Type, Status]) Run(ctx context.Context) {
 	// Ensure that the background consumers are only initialized once
 	w.once.Do(func() {
+		ctx, cancel := context.WithCancel(ctx)
+		w.cancel = cancel
 		w.calledRun = true
 
-		for currentStatus, processes := range w.processes {
-			for _, p := range processes {
+		for currentStatus, consumers := range w.consumers {
+			for _, p := range consumers {
 				if p.ParallelCount < 2 {
-					// Launch all processes in runners
+					// Launch all consumers in runners
 					go consumer(ctx, w, currentStatus, p, 1, 1)
 				} else {
 					// Run as sharded parallel consumers
@@ -277,246 +98,53 @@ func (w *Workflow[Type, Status]) Run(ctx context.Context) {
 	})
 }
 
-func consumer[Type any, Status ~string](ctx context.Context, w *Workflow[Type, Status], currentStatus Status, p process[Type, Status], shard, totalShards int) {
-	if w.debugMode {
-		log.Info(ctx, "launched consumer", j.MKV{
-			"workflow_name": w.Name,
-			"from":          currentStatus,
-			"to":            p.DestinationStatus,
-		})
-	}
-
-	role := makeRole(
-		w.Name,
-		string(currentStatus),
-		"to",
-		string(p.DestinationStatus),
-		"consumer",
-		fmt.Sprintf("%v", shard),
-		"of",
-		fmt.Sprintf("%v", totalShards),
-	)
+// Stop cancels the context provided to all the background processes that the workflow launched and waits for all of
+// them to shut down gracefully.
+func (w *Workflow[Type, Status]) Stop() {
+	// Cancel the parent context of the workflow to gracefully shutdown.
+	w.cancel()
 
 	for {
-		ctx, cancel, err := w.scheduler.Await(ctx, role)
-		if err != nil {
-			log.Error(ctx, errors.Wrap(err, "consumer error"))
-		}
-
-		if w.debugMode {
-			log.Info(ctx, "consumer obtained role", j.MKV{
-				"role": role,
-			})
-		}
-
-		if ctx.Err() != nil {
-			// Gracefully exit when context has been cancelled
-			if w.debugMode {
-				log.Info(ctx, "shutting down process - consumer", j.MKV{
-					"workflow_name":      w.Name,
-					"current_status":     currentStatus,
-					"destination_status": p.DestinationStatus,
-					"shard":              shard,
-					"total_shards":       totalShards,
-					"role":               role,
-				})
+		var runningProcesses int
+		for _, state := range w.States() {
+			switch state {
+			case StateUnknown, StateShutdown:
+				continue
+			default:
+				runningProcesses++
 			}
+		}
+
+		// Once all processes have exited then return
+		if runningProcesses == 0 {
 			return
 		}
-
-		err = runStepConsumerForever[Type, Status](ctx, w, p, currentStatus, role, shard, totalShards)
-		if errors.IsAny(err, ErrWorkflowShutdown, context.Canceled) {
-			if w.debugMode {
-				log.Info(ctx, "shutting down process - consumer", j.MKV{
-					"workflow_name":      w.Name,
-					"current_status":     currentStatus,
-					"destination_status": p.DestinationStatus,
-				})
-			}
-		} else if err != nil {
-			log.Error(ctx, errors.Wrap(err, "consumer error"))
-		}
-
-		select {
-		case <-ctx.Done():
-			cancel()
-			if w.debugMode {
-				log.Info(ctx, "shutting down process - consumer", j.MKV{
-					"workflow_name":      w.Name,
-					"current_status":     currentStatus,
-					"destination_status": p.DestinationStatus,
-				})
-			}
-			return
-		case <-time.After(p.ErrBackOff):
-			cancel()
-		}
 	}
 }
 
-func timeoutPoller[Type any, Status ~string](ctx context.Context, w *Workflow[Type, Status], status Status, timeouts timeouts[Type, Status]) {
-	if w.debugMode {
-		log.Info(ctx, "launched timeout consumer", j.MKV{
-			"workflow_name": w.Name,
-			"for":           status,
-		})
+func update(ctx context.Context, streamer EventStreamer, store RecordStore, wr *WireRecord) error {
+	body, err := wr.ProtoMarshal()
+	if err != nil {
+		return err
 	}
 
-	role := makeRole(w.Name, string(status), "timeout-consumer")
-
-	for {
-		ctx, cancel, err := w.scheduler.Await(ctx, role)
-		if err != nil {
-			log.Error(ctx, errors.Wrap(err, "timeout auto inserter consumer error"))
-		}
-
-		err = pollTimeouts(ctx, w, status, timeouts)
-		if errors.IsAny(err, ErrWorkflowShutdown, context.Canceled) {
-			cancel()
-			return
-		} else if err != nil {
-			log.Error(ctx, errors.Wrap(err, "timeout consumer error"))
-		}
-
-		select {
-		case <-ctx.Done():
-			cancel()
-			return
-		case <-time.After(timeouts.ErrBackOff):
-			cancel()
-		}
-	}
-}
-
-func timeoutAutoInserterConsumer[Type any, Status ~string](ctx context.Context, w *Workflow[Type, Status], status Status, timeouts timeouts[Type, Status]) {
-	if w.debugMode {
-		log.Info(ctx, "launched timeout auto inserter consumer", j.MKV{
-			"workflow_name": w.Name,
-			"for":           status,
-		})
+	e := Event{
+		ForeignID: wr.ForeignID,
+		Body:      body,
+		Headers:   make(map[string]string),
 	}
 
-	role := makeRole(w.Name, string(status), "timeout-auto-inserter-consumer")
-
-	for {
-		ctx, cancel, err := w.scheduler.Await(ctx, role)
-		if err != nil {
-			log.Error(ctx, errors.Wrap(err, "timeout auto inserter consumer error"))
-		}
-
-		cunsumerFunc := func(ctx context.Context, r *Record[Type, Status]) (bool, error) {
-			for _, config := range timeouts.Transitions {
-				ok, expireAt, err := config.TimerFunc(ctx, r, w.clock.Now())
-				if err != nil {
-					return false, err
-				}
-
-				if !ok {
-					// Ignore and evaluate the next
-					continue
-				}
-
-				err = w.timeoutStore.Create(ctx, r.WorkflowName, r.ForeignID, r.RunID, string(status), expireAt)
-				if err != nil {
-					return false, err
-				}
-			}
-
-			// Never update state even when successful
-			return false, nil
-		}
-
-		err = runStepConsumerForever(ctx, w, process[Type, Status]{
-			PollingFrequency: timeouts.PollingFrequency,
-			ErrBackOff:       timeouts.ErrBackOff,
-			Consumer:         cunsumerFunc,
-			ParallelCount:    1,
-		},
-			status,
-			role,
-			1,
-			1,
-		)
-		if errors.IsAny(err, ErrWorkflowShutdown, context.Canceled) {
-			if w.debugMode {
-				log.Info(ctx, "shutting down process - timeout auto inserter consumer", j.MKV{
-					"workflow_name": w.Name,
-					"status":        status,
-				})
-			}
-		} else if err != nil {
-			log.Error(ctx, errors.Wrap(err, "timeout auto inserter consumer error"))
-		}
-
-		select {
-		case <-ctx.Done():
-			cancel()
-			return
-		case <-time.After(timeouts.ErrBackOff):
-			cancel()
-		}
+	err = store.Store(ctx, wr)
+	if err != nil {
+		return err
 	}
-}
 
-func makeRole(inputs ...string) string {
-	joined := strings.Join(inputs, "-")
-	lowered := strings.ToLower(joined)
-	filled := strings.Replace(lowered, " ", "_", -1)
-	return filled
-}
-
-type process[Type any, Status ~string] struct {
-	PollingFrequency  time.Duration
-	ErrBackOff        time.Duration
-	DestinationStatus Status
-	Consumer          ConsumerFunc[Type, Status]
-	ParallelCount     int
-}
-
-type callback[Type any, Status ~string] struct {
-	DestinationStatus Status
-	CallbackFunc      CallbackFunc[Type, Status]
-}
-
-type timeouts[Type any, Status ~string] struct {
-	PollingFrequency time.Duration
-	ErrBackOff       time.Duration
-	Transitions      []timeout[Type, Status]
-}
-
-type timeout[Type any, Status ~string] struct {
-	DestinationStatus Status
-	TimerFunc         TimerFunc[Type, Status]
-	TimeoutFunc       TimeoutFunc[Type, Status]
-}
-
-type ConsumerFunc[Type any, Status ~string] func(ctx context.Context, r *Record[Type, Status]) (bool, error)
-
-type CallbackFunc[Type any, Status ~string] func(ctx context.Context, r *Record[Type, Status], reader io.Reader) (bool, error)
-
-type TimerFunc[Type any, Status ~string] func(ctx context.Context, r *Record[Type, Status], now time.Time) (bool, time.Time, error)
-
-type TimeoutFunc[Type any, Status ~string] func(ctx context.Context, r *Record[Type, Status], now time.Time) (bool, error)
-
-func Not[Type any, Status ~string](c ConsumerFunc[Type, Status]) ConsumerFunc[Type, Status] {
-	return func(ctx context.Context, r *Record[Type, Status]) (bool, error) {
-		pass, err := c(ctx, r)
-		if err != nil {
-			return false, err
-		}
-
-		return !pass, nil
+	topic := Topic(wr.WorkflowName, wr.Status)
+	producer := streamer.NewProducer(topic)
+	err = producer.Send(ctx, &e)
+	if err != nil {
+		return err
 	}
-}
 
-func DurationTimerFunc[Type any, Status ~string](duration time.Duration) TimerFunc[Type, Status] {
-	return func(ctx context.Context, r *Record[Type, Status], now time.Time) (bool, time.Time, error) {
-		return true, now.Add(duration), nil
-	}
-}
-
-func TimeTimerFunc[Type any, Status ~string](t time.Time) TimerFunc[Type, Status] {
-	return func(ctx context.Context, r *Record[Type, Status], now time.Time) (bool, time.Time, error) {
-		return true, t, nil
-	}
+	return producer.Close()
 }
